Use le-automaton token for editor copy update
We set `GITHUB_TOKEN` to be read-only at the `abetterinternet` org
level, but this breaks the "Update GitHub Pages" job in the "Update
Editor's Copy" workflow. We now instead use a [personal account
token](https://docs.github.com/en/github/authenticating-to-github/keeping-your-account-and-data-secure/creating-a-personal-access-token)
generated from the `le-automaton` GitHub account which is scoped to
allow actions on private repos.

Additionally, this deletes the `archive` workflow which I don't think we
need: GitHub itself is a perfectly good durable store of issues and PRs,
and updating GH pages on push to main should suffice.