Amend PDUs to support multiple helpers
Modify `struct Param`, `struct Report` and `struct CollectResp` to
accommodate more than one helper. Note that the PPM protocol doesn't
actually support multiple helpers to enhance privacy (#68) but we
nonetheless want to define the PDUs so it is possible to introduce that
feature in the future. We do this by introducing an `AggregatorId` type
into various struct definitions, allowing participants to tell which
aggregator an input share or output share corresponds to. We also
introduce a special, reserved value of `AggregatorId` that allows
clients to tell which aggregator is the leader to which reports should
be uploaded. Finally, we use the aggregator ID in HPKE context
construction instead of the existing constants 0x00 and 0x01 for leader
and helper, respectively.

Along the way, we introduce the `HpkeConfigId` type alias used in few
different places.

Resolves #117, #133