Amend PDUs to support multiple helpers
Modify `struct Param`, `struct Report` and `struct CollectResp` to
accommodate more than one helper. Note that the PPM protocol doesn't
actually support multiple helpers to enhance privacy (#68) but we
nonetheless want to define the PDUs so it is possible to introduce that
feature in the future.

We also move `task_id` from the HPKE AAD to the context's application
info, and introduce the `HpkeConfigId` type alias in a few different
places.

Resolves #117, #133