Resolve open batch boundary issues
We have the following security/operational considerations, which we
can't currently address at the same time.

  - Enforcing privacy budgets for DP requires us to limit the number of
  batches the same report is used.
  - Hits requires the same report to be used in some number of batches.

To resolve this, this change adds a `max_batch_lifetime` parameter,
which specifies the maximum number of times a batch of reports may be
used in a collect request.

This change also seeks to clarify some ambiguities in how the batch
boundaries are defined, and why they're defined that way. Along the way,
it cleans up some protocol bits that got lost in the last refactor.