Clarify timing of report aggregations & purpose of anti-replay.
As previously written, the timing of report aggregations would allow
aggregation of reports with timestamps far into the future, which would
allow a single client reporting a timestamp of INT_MAX to DOS the
aggregation system due to the way timestamps are used in the anti-replay
mechanism. Given the severity of not following this portion of the
specification, I upgrade the SHOULD to a MUST as well.