Change grace_window behavior from a MUST to a SHOULD.
With the previous changes to the following MUST NOT, a SHOULD makes more
sense. A suitably pathological implementation will still see some funky
behavior[1], but it's not so bad that another MUST is required IMO.

[1] Specifically, an implementation that chooses to ignore the
grace_window SHOULD, and accepts reports clock-skewed a few minutes in
the future, could end up dropping non-skewed reports if a skewed report
arrives just before an aggregate request is made.