Change anti-replay mechanism
Instead of requiring reports to be processed in order, require only that
nonces are not replayed. This means that the aggregators will need to
keep track of the nonces of processed reports. The upside is that the
leader no longer needs to re-order out-of-order reports.

Another change we must make is to ignore reports that arrive after their
batch interval has been collected. Otherwise the aggregate result over
the new batch may leak the measurement.

Co-authored-by: Tim Geoghegan <timg@letsencrypt.org>
Co-authored-by: Christopher Wood <caw@heapingbits.net>