DO NOT MERGE Proposal for first interop target
The current version of the protocol cannot be implemented in the sense
that it does not specify a mapping from the VDAF execution to HTTP
requests. The goal of this PR is to flesh out one possible mapping so
that we can begin work on implementation, with the understanding of
course that the overall shape and the details are still subject to
change.

The design presumes we have reached consensus on a variety of open
issues, all of which still need to be discussed. The purpose of this PR
is *not* to facilitate these discussions, but instead to ensure sure
there is a solution to each issue that we're confident we can implement.
These issues include:

* PR #169 will be merged, i.e., the anti-replay mechanism will be
relaxed to require the helper to store nonce sets.

* PR #174 will not be merged, i.e., clients will continue to upload
reports to the leader (and not the helpers). However, if we end up
choosing to take this PR, it shouldn't be too bad to change course.

* Issue #45 will be resolved by adding support for VDAFs w/o
verifiability. (We should probably call these "DAFs".)

* Issue #68 will be resolved by adding support for multiple helpers (for
applicable VDAFs). We don't do this quite yet, but the aggregate flow is
meant to be easily extendable to accommodate this.

* Messages sent between aggregators need to be authenticated.

We have an answer for each of these, but we will discuss and merge PRs
for each of them separately.

**Tips for reviewing:** Rather than look at the diff, I would read the
document from /Work begins here/ to /Work ends here/. There is a big new
section in the middle, and GH didn't do a very good job of showing the
change.