Align some terminology with the VDAF draft
What PPM calls an "output share" is called an "aggregate share" in the
VDAF draft. Furthermore, in the VDAF draft, an "output share" is what
an aggregator recovers from running the MPC with its peers. This change
fixes this ambiguity without changing the PPM protocol itself (modulo
changing the info string for HPKE).