Redefine `Interval` to invalid values impossible
While reviewing #179, @cjpatton and I figured out a way to redefine
`struct Interval` so that invalid values cannot be represented[1].

The existing representation contains `Time start` and `Time end`, which
means an interval could be defined with `end < start`, which is invalid.
Redefining `Interval` as a start instant and a duration makes it
impossible to represent invalid `Interval`s (since `Duration` is an
unsigned integer).

Additionally, this implicitly fixes a bug in the batch interval validation
logic: we asserted that `batch_interval.end` should be divisible by
`min_batch_interval`, but since `end` is _excluded_ from the `Interval`, I
think that was incorrect. Instead, `batch_interval.end - 1` should be divisible
by `min_batch_duration`. In any case, this is simpler with the new
definition of `Interval`: `Interval.duration` needs to be a multiple of
`min_batch_duration`.

[1]: https://github.com/abetterinternet/ppm-specification/pull/179#discussion_r789340693