fix HPKE info strings and specify agg share handling
The info strings used in HPKE context construction used the prefix
"pda", referring to the old "Private Data Aggregation" name of this
protocol. This commit also adds text instructing the collector how to
decrypt the aggregate shares it receives and to use the VDAF's
unsharding algorithm to obtain aggregate results.