add a `task_id` query parameter for `hpke_config`
All protocol request messages contain a `task_id` field, which allows
aggregators to service multiple tasks from a single set of endpoints.
The exception is `hpke_config`: since it is an HTTP GET request, there
is no body, and thus no place for the client to indicate the desired
`task_id` to the aggregator. We don't want to introduce a request body
to `hpke_config`, because then it would have to be a POST instead of a
GET, and that would make it harder for implementations to cache HPKE
config values. So instead we have clients encode the task ID in RFC 4648
base64url without padding and provide it as a query parameter.

See issue #146