Add VDAF mapping for aggregation flow
This wires up PPM to drive the underlying VDAF state machine for
individual reports. It splits the aggregate flow into three phases:
initialization, progression (where the aggregators exchange messages
to perform validation), and finalization. It also moves the process
of fetching the aggregate share value for each aggregator to the
collect flow, since that seems more closely aligned with the process
of collection than it is about aggregation.