Add support for fixed-size tasks
Generalizes "batch_interval" to "query" and enumerates the different
query types and constraints that apply to each. The existing collect
request semantics is subsumed by the `time-interval` query type.

Adds a query type, `fixed-size`, that is designed to support use cases
in which the sample size needs to be tightly controlled. The main
feature of this query type is that the Aggregators enforce limits on the
size of each batch. We given them a little wiggle room in order to make
coordination of aggregation jobs a bit easier.

One notable feature of this change is that each tasks supports exactly
one query type. This allows us to punt on the coomplexity of having to
composing different query types per task. For example: If a report can
be batched into a chunk or a batch interval, then it's not clear that
it's possible for Aggregators to aggregate reports ahead of a collect
request.

Finally, a new struct is defined, `ReportMetadata`, that subsumes all
the per-`Report` information required for aggregation. This change is
mainly editorial: It allows us to more cleanly spell out the AAD passed
to HPKE encryption, but does not change the wire format.

Co-authored-by: Tim Geoghegan <timg@letsencrypt.org>
Co-authored-by: Brandon Pitman <bran@bran.land>
Co-authored-by: Christopher Wood <caw@heapingbits.net>