Merge "fixed-interval" & "fixed-size" task types (mostly).
Almost all functionality & message wire formats are the same for the two
task types again. The largest point of divergence is in the batch
validation implemented by the leader & helper: this validation is used
to enforce the desired properties on each batch (having a fixed interval
of time or including a fixed number of reports, respectively).