Clarify handling of replayed reports in "Upload Requests"
We didn't have any text that actually instructed leaders to reject
uploads deemed to be replays based on their metadata. We now have a
paragraph explaining that leaders MUST ignore such reports and MAY alert
the client. This allows leader implementations to defer this check to
the aggregation phase as described in {{input-share-validation}}.

We define a new problem document type `reportRejected`, which replaces
`reportTooLate`, previously used in the case where the leader detects a
report pertaining to an already-collected batch. The rationale for using
a less specific error type for these two cases is twofold:

  - In these cases, we don't expect the client to do anything but give up
    on the report, so nothing more specific than a generic rejection is
    needed;
  - If we did provide more specific error codes, we'd be giving
    attackers an oracle for what nonces have already been uploaded or
    what reports have already been collected.

While I was in this text, I also:

  - clarified what the client should do if it gets `reportTooEarly`. If
    the client shouldn't do anything special in that situation, then I
    argue the error should be the generic `reportRejected`.
  - downgraded a MUST to a SHOULD in the case where the client gets
    `outdatedConfig` twice. There's no way the protocol can meaningfully
    enforce that MUST, and if the client continues to upload reports
    referencing an outdated or non-existent HPKE config, the leader can
    just continue to ignore the reports.

Resolves #191