Allow per-report artificats to be dropped before task expiration
Closes #180.

Clarify that `report-dropped` is used when it's impossible to determine
if an input share is valid. One situation we need this for is when an
Aggregator drops some anti-replay state in order to reduce storage
costs. This should result in data loss in the worst case, not a
violation of privacy.