Ensure all protocol messages are self-describing
Currently parsing AggreateInitializeReq, AggregateshareReq, CollectReq,
and CollectResp requires the query type associated with the task. The
query type is determined by the task, but it is necessary to parse the
task ID from the message before the remainder can be parsed.

This change ensures that parsing of these messages does not require the
query type.