Editorial: Define PartialBatchSelector
The BatchSelector is the value agreed upon by the Aggregators during the
collect flow. It is used to guide selection of a batch of reports for
computing the aggregate result.

Conceptually, a PartialBatchSelector is "promoted" to a BatchSelector
once the Collector's query is determined. The PartialBatchSelector is
used during the aggregation flow to assign reports to batches. For
example, for fixed size queries, the BatchID is selected during the
Leader during the aggregation flow.