DO NOT MERGE sketch of resource oriented DAP API
This is a sketch of a resource oriented API for DAP. It's presented
separately from the main body of protocol text because before taking on
the work of integrating it in, I want to begin a discussion about this
proposal, leading up to a presentation at IETF 115.

The objectives are to refactor the DAP API into a resource-oriented API
more in line with conventional HTTP semantics and REST. In particular, I
am trying to pay close attention to idempotence and ensuring that
protocol participants can recover gracefully from network messages being
lost.

Designing the API around resources identified by a URI also simplifies
message handling: servers no longer need to parse the first part of a
message to get a task ID to then figure out what the rest of the message
contains, because now things like a task ID are in a resource's path.

We also should be able to write less protocol text, because if all of
our resources are boring HTTP nouns, then the usual semantics for things
like HTTP 202 Accepted, 3xx redirects or 404 Not Found apply.

I do _not_ introduce an ACME style API directory, because I'm not
convinced that it is appropriate to introduce an extra level of
indirection for DAP API requests, which are going to be more performance
sensitive than ACME.

I also deliberately avoid ACME's POST-as-GET requests, because our
experience at Let's Encrypt has been that those make it hard to cache
responses, because CDNs don't expect POST results to be cacheable.

I've put this up on GitHub as a PR so that it's out in public and folks
can leave comments inline reasonably gracefully, although I don't intend
to ever merge this.

Relevant to #278