Make aggregation job URI opaque
Move the responsibility of constructing aggregation job resource URIs to
the helper, to bring this resource in line with how collections and
aggregate shares are handled. We also remove the `GET` method
requirement for aggregation jobs, since [the sequence diagram in this
comment][1] illustrates how a leader can recover from a crash during the
aggregate sub-protocol without querying the helper for job state.
Finally, we unify `AggregateJobPutReq` into `AggregateJob`, which makes
some messages in the aggregate sub-protocol bigger but leaves us with
fewer message definitions overall.

[1]: https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/pull/367#issuecomment-1306289706