Introduce resource oriented API
Changes the DAP API to orient the upload, aggregate and collect
sub-protocols around some HTTP resources, namely:

Upload:
 - HPKE configurations
 - Reports

Aggregate:
 - Aggregation jobs

Collect:
 - Collections
 - Aggregate shares

There's a lot of change here, most of which is changing the spelling of
HTTP endpoints from e.g. `POST /upload` to
`PUT /tasks/{task-id}/reports/{report-id}`. But there are some more
important improvements made here:

 - move some identifiers from message bodies into resource paths,
   removing the need to partially parse messages to get task ID
 - makes Helper aggregate share handling asynchronous, mirroring
   collection handling Leader
 - aggregation job messages now contain a `round` field to allow
   aggregators to recover from message loss

The other big goal here is to make better use of HTTP primitives and
conventions. However, it's important not to get too precious about this:
we expect that anything using this API is a DAP implementation rather
than a generic HTTP client like a web browser, so it doesn't matter that
much if we choose exactly the right HTTP method, because ultimately DAP
gets to dictate the semantics of any message.