Add `interval` to struct `Collection`
The Leader now informs the Collector of what time interval is spanned by
the reports aggregated into a collection, regardless of the query type.
This is particularly useful for fixed size tasks as otherwise a
Collector can only guess at the report times based on the timing of
collection requests, but also useful in time interval tasks since the
aggregated reports could span a *smaller* interval that what was
requested (e.g., if the collector requested a one hour time interval,
but reports only arrived during minutes 30-40 of that hour).

Resolves #397