edit: aggregation nits and distinguish prep_msg from prep_share
Continue editorial pass up to the collection flow. Along, clarify an
ambiguity in the mapping of VDAF preparation to DAP aggregation
initialization/continuation.

In particular, `VDAF.prep_next()` is meant to take in the "preperation
message" from the previous round and output either (1) the Aggregator's
output share or (2) the Aggregator's next preparation sate and next
"preparation message-share".

We are currently using the term "inbound message" for the former and
"outbound message" for the latter. This seems to be causing us to
conflate "preparation message" with "preparation message-share", which
may result in misunderstanding the protocol flow.