Don't RECOMMEND how to generate IDs
The current recommendation is to generate it using a "cryptographically
secure random number generator", which is usually reserved for purposes
that need a uniform random value. All we need is uniqueness; the purpose
of this text is merely to suggest a convenient way of meeting the
requirement.

To avoid confusion, just drop the recommendation altogether.

Co-authored-by: Tim Geoghegan <timg@letsencrypt.org>