Encapsulate details of VDAF preparation into "prep steps"
The VDAF spec distinguishes between two types of intermediate values
computed and exchanged amongst the Aggregators: "prep shares" and "prep
messages". This distinction is largely irrelevant in the current
aggregation flow (i.e., that of draft04), since the Leader always sends
the prep message to the Helper and the Helper always sends its prep
share to the Leader.

In the proposed "ping-pong" flow, the distinction is important because
the Leader and Helper take turns sending these values. As a result,
whether the host is processing a prep share or prep message depends on
which step of preparation is being computed.

In PR#393 (by @branlwyd), these details are encapsulated by the
`PrepareStep` message, which defines a variant for each combination of
values encountered during the aggregation flow. In order to simplify the
overview, this change hoists this encapsulation up to the overview
section, thereby removing the low details that are irrelevant to the
overall structure.

While at it, this change defines two procedures, `VdafPrepInit()` and
`VdafPrepNext()`, that specify explicitly the prep step processing
rules. The goal is to make the description of the aggregation flow in
the following subsections more modular and easy to follow.