Remove queryMismatch, in favor of unrecognizedMessage.
The aggregator behavior on receiving this error is the same either way,
and there are many other message-based errors that do not get their own
code. This change effectively leaves it up to implementations to decide
the level of detail they include in error reports.

unrecognizedMessage might seem like a strange error to use since the
message is invalid, not unrecognized; however, this error is used for
all message-parsing errors, and may soon be renamed to invalidMessage
(in #464).