Specify one Helper & compute two VDAF rounds per DAP aggregation step.
This is based on the "ping-pong" communication pattern described in the
VDAF specification, and reduces the number of network round-trips
required to complete an aggregation by about half for all VDAFs, from
`ROUNDS + 1` to `ceil((ROUNDS + 1) / 2)`. In particular, for 1-round
VDAFs like those in the Prio family, this reduces aggregation to a
single network round-trip.

Implementing this change requires specifying to exactly two aggregators,
i.e. exactly one Helper. The text of the aggregation section is updated
to explicitly specify one Helper.