Rename ReportShareError to PrepareError.
The old name is confusing to me because almost none of the errors
represented by this type are directly related to the report share. I
chose the name PrepareError because this error is used during the
report preparation (aka aggregation) process, matching other types
similarly used during the report preparation process.