Defer leader batch validation
Move references to `{{batch-validation}}` in the leader's text to just
before it requests an aggregate share from the helper, since validating
the batch at the time of creating a collection job is premature. We do
however allow for leader implementations to implement batch validation
checks earlier if they want to.

Refine the text in `{{batch-validation}}` to capture some error handling
nuances: the helper should always abort and do so with a synchronous
response to the `AggregateShareReq` it got, but the leader's errors will
be transmitted in subsequent responses to `POST` on a collection, and
also too few valid reports isn't a fatal error.

Resolves #466