Clarify max_batch_query_count is counted by distinct aggregation params.
I believe this is the correct check, as re-collecting a batch against an
already-collected aggregation parameter will not reveal additional
information -- it is the number of distinct aggregation parameters that
determiens the number of different collections that have been made
against a given batch.

See https://github.com/divviup/janus/pull/1799 for context.