Clarify mapping of input shares to leader/helper, encoding requirements.
Specifically, call out that the sharding algorithm will return two input
shares, and the first one is the Leader's.

Also describe that DAP only works with VDAFs which have defined
encodings for the various VDAF types which DAP embeds into its own types
as opaque byte strings. I purposefully leave the set of messages which
require a defined encoding ambiguous in hopes that this avoids the
possibility of forgetting to update the list as the specification is
changed.