Specify aggregation with batched preparation
Adds an alternative aggregation flow for VDAFs that support batched
preparation.

In this mode of operation, prep shares for many reports are aggregated
into a short message called a "tag". The tags computed by the Leader and
Helper are used to decide validity of the reports: if the tags are
valid, then every report is deemed valid; otherwise, at least one report
is known to be invalid.

The Helper handles invalidity by aborting the request. In this case the
Leader is allowed to retry, either with a subset of the reports or with
the existing, pep-report preparation code path.