Rename "query type" to "batch mode", "fixed-size" to "leader-selected", remove max_batch_size.
Batch mode is a more descriptive name than query type. And
leader-selected is a more accurate name than fixed-size, since the batch
sizes are not fixed -- any size above min_batch_size (and, if specified,
below max_batch_size) is acceptable.

The max_batch_size parameter of the leader-selected query type is
removed. This parameter does not need to be enforced by both
aggregators, so it can be made implementation-specific.

Also, restore the partial batch selector to the Collection message; the
included batch ID is required for the Collector to decrypt the aggregate
shares, as it is included in the AAD. This bug was introduced in #554 &
effectively breaks collection for the fixed-size query type in DAP-11.