Simplify report replay checking.
Specifically, we relax the check from the specific report ID, to
intersection with intervals of report IDs. This will allow
implementations to drop all long-term per-report storage, considerably
reducing the storage requirements to implement the protocol. The
tradeoff is in increased aggregation latency: typically, aggregation
cannot begin until the Leader is sure that no further reports will
arrive for a given time window.