Document more storage and runtime requirements
Besides the size of the underlying measurements, the runtime may vary
based on the aggregation parameter. We motivate this with Poplar1 here,
but future VDAFs, e.g., draft-mouris-cfrg-mastic, may provide more
pertinent motivation.

It is also worth noting that storage requirements vary by the size of
the underlying measurements and/or aggregation parameter.