Update PrepareError allocations (*)
Wherever we have extension points in the protocol, we should reserve a
`0` for testing purposes. We've done this for upload extensions and
batch modes, but failed to do so for prepare errors. This is relevant
because future DAP extensions might need to register prepare errors.

Reserve `0` and shift all remaining codepoints by `1`, except for those
following `6`, which was allocated in a previous draft but earmarked for
deletion in a future draft.