Rename "prepare error" to "report error"
The term "prepare" refers to "VDAF preparation". Yet PrepareError
enumerates all reasons a report might be rejected during aggregation,
just one of which has to do with VDAF preparation. The others are
strictly about DAP protocol logic.

Rename to "report error", as this is evokes an error that prevents a
report from being aggregated.