Aggregation step skew recovery: restart from beginning.
This replaces the existing aggregation step skew recovery scheme, which
restarts from the previous step.

This likely simplifies the Helper, which no longer needs to remember the
response it sends to each aggregation request. On the other hand, this
somewhat complicates the Leader, which must now be able to replay the
initial request, and which must understand when it needs to restart. The
Leader is no longer expected to durably store each step of the
aggregation process, however.