Clarify rejection criterion for future reports
The spec encourages rejection of reports with timestamps too far in the
future. The Aggregator's action in this case is unspecified. It says if
a report is rejected because it's too searly, then the Aggregator
"SHOULD mark an input share rejected with error `report_too_early`".
This allows for the possibility of a different error.

Co-authored-by: Brandon Pitman <bran@bran.land>