Enforce agg param validity (*)
Before aggregating a batch, it is necessary to confirm that the
aggregation parameter is valid by decoding it and running `is_valid()`.

While at it, remove the `batchQueriedMultipleTimes` abort type, since
the condition it's used in is no longer possible.

Note that this is a breaking change, since previously the aggregation
parameter would have been ignored.