Correct usage of URN, URI, URL
URNs (RFC 2141) are of the form `urn:...`.

URIs uniquely identify a resource, e.g. `example.com/my-document`. A URL
is a URI that includes information necessary to to access the identified
resource, such as the `https` scheme (RFC 3986).

Anywhere we're discussing problem types, we should be saying URN.
Otherwise this document discusses resources served at URIs including a
scheme, which are URLs, so we should always say that.