tweak {#batch-mode} text
Moves some text and definitions from the "batch modes" section in the
overview to the top-level "batch modes" section. The intent is to trim
what's in the overview down to the ideas necessary to parse the rest of
the document, and keep all the normative text about batch modes (e.g.,
enum values) in one place.