tweak aggregation job text to align with collect
Review of the changes to collection job specification (#675) yielded
some tweaks we now apply back to aggregation jobs: we introduce a notion
of a job _failing_ that is valid regardless of whether an error occurs
during synchronous or asynchronous. Previously, we discussed the helper
"aborting", which only makes sense if there is a leader waiting for a
response to an HTTP request. Finally, we make it clear when aggregation
job responses are considered "ready".