Clarify batch validation for leader
The Leader's validation of collection jobs is a little fuzzy as
discussed in #662. No small part of the problem is that we refer to a
single {#batch-validation} section from both the Leader's text on
collection jobs and the Helper's text on aggregate share jobs. This
commit resolves the Leader-side problems in #662 by folding
{#batch-validation} into {#collect-init}. We also address ambiguities
about which of the Leader or Collector can retry failing collection
jobs.

- Collector may retry collection jobs if they fail with
  `invalidBatchSize`
- Leader checks if query is valid. Update the reference for query
  validity to {#batch-modes} instead of {#batch-buckets}, since only the
  former discusses validity of queries or partial batch selectors.
- Leader checks for batch overlap
- Leader checks if enough validated reports are available after running
  relevant aggregation jobs to completion.

An important nuance of the last point: who is responsible for retrying
collection jobs if there aren't enough reports available yet? With this
text, it can be either the Leader or the Collector.

The Leader may retry collection jobs and keep the Collector waiting,
either blocking its response (synchronous handling case) or indicating
the job is not ready yet (asynchronous handling case). The Leader may
also fail right away with `invalidBatchSize` (it can also do this after
however many retries it likes). In that case, the Collector can retry
the collection job later.

In either case, the idea is that they can wait until more reports get
uploaded and validated, but DAP doesn't spell out a mechanism or
heuristic for that. This does admit the possibility of "double
retries", where the Leader might retry a few times, then give up with
`invalidBatchSize`, and then the Collector might retry a few more
times. Improving this would require some more protocol engineering
(perhaps you'd need the `invalidBatchSize` error constructed by the
Leader to include how many times it tried and how much wall time it
allowed to elapse before giving up, so the Collector could decide if it
wants to retry).