Clarify batch validation for Helper
The Helper side of clarifications informed by #662. The text from
{#batch-validation} is now folded into {#collect-aggregate}. Besides
moving text around, we also remove the prescription that when the Leader
gets `invalidBatchSize` from the Helper, the Leader should retry the
request later. This should never happen, because the Leader has to run
all relevant aggregation jobs to completion before sending an aggregate
share request, and so should know exactly how many validated reports
there are in the batch.